+++ src/polyfills/symbol.async-iterator.ts
@@ -1,0 +1,9 @@
+interface SymbolConstructor {
+    readonly asyncIterator: symbol;
+}
+
+if (!Symbol.asyncIterator) {
+    Object.defineProperty(Symbol, 'asyncIterator', {
+        value: Symbol('Symbol.asyncIterator'),
+    });
+}