+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## [Unreleased]
 
+### Internal API
+
+* [#48] - Fetch Netlify API only when needed
+
 [Unreleased]: https://github.com/sounisi5011/metalsmith-netlify-published-date/compare/v0.1.0...HEAD
+[#48]: https://github.com/sounisi5011/metalsmith-netlify-published-date/pull/48
 
 ## [0.1.0] (2019-08-07)